+++ tutor/templates/apps/openedx/settings/partials/common_all.py
@@ -32,6 +32,13 @@ for store in MODULESTORE["default"]["OPTIONS"]["stores"]:
 # Behave like memcache when it comes to connection errors
 DJANGO_REDIS_IGNORE_EXCEPTIONS = True
 
+# Elasticsearch connection parameters
+ELASTIC_SEARCH_CONFIG = [{
+  {% if ELASTICSEARCH_SCHEME == "https" %}"use_ssl": True,{% endif %}
+  "host": "{{ ELASTICSEARCH_HOST }}",
+  "port": {{ ELASTICSEARCH_PORT }},
+}]
+
 DEFAULT_FROM_EMAIL = ENV_TOKENS.get("DEFAULT_FROM_EMAIL", ENV_TOKENS["CONTACT_EMAIL"])
 DEFAULT_FEEDBACK_EMAIL = ENV_TOKENS.get("DEFAULT_FEEDBACK_EMAIL", ENV_TOKENS["CONTACT_EMAIL"])
 SERVER_EMAIL = ENV_TOKENS.get("SERVER_EMAIL", ENV_TOKENS["CONTACT_EMAIL"])
@@ -86,6 +93,11 @@ LOGGING["handlers"]["tracking"] = {
     "formatter": "standard",
 }
 LOGGING["loggers"]["tracking"]["handlers"] = ["console", "local", "tracking"]
+# Silence some loggers (note: we must attempt to get rid of these when upgrading from one release to the next)
+import warnings
+warnings.filterwarnings("ignore", category=DeprecationWarning, module="newrelic.console")
+warnings.filterwarnings("ignore", category=DeprecationWarning, module="lms.djangoapps.course_wiki.plugins.markdownedx.wiki_plugin")
+warnings.filterwarnings("ignore", category=DeprecationWarning, module="wiki.plugins.links.wiki_plugin")
 
 # Email
 EMAIL_USE_SSL = {{ SMTP_USE_SSL }}